Fix: Change rating type from Int to Float in Review schema
- Change rating field from Int to Float in Review type to support decimal ratings (e.g., 4.5)
- Change rating field from Int! to Float! in ReviewInput to match
- This fixes the error: 'Int cannot represent non-integer value: 4.5'

The database model already stores rating as Number (supports decimals), but the GraphQL schema was incorrectly defined as Int.

+++ src/graphql/schema.js
@@ -69,7 +69,7 @@ const typeDefs = gql`
 
   type Review {
     _id: ID
-    rating: Int
+    rating: Float
     description: String
     isActive: Boolean
     createdAt: Date
@@ -1129,7 +1129,7 @@ const typeDefs = gql`
 
   input ReviewInput {
     order: String!
-    rating: Int!
+    rating: Float!
     description: String
   }
